+++ example/modifier_volume_with_two_volumes.cpp
@@ -37,8 +37,8 @@ static ft0::data_transfer::data_format data[4];
 
 void do_get_condition()
 {
-  uint32_t send_buf[1024] __attribute__((aligned(32)));
-  uint32_t recv_buf[1024] __attribute__((aligned(32)));
+  uint8_t send_buf[1024] __attribute__((aligned(32)));
+  uint8_t recv_buf[1024] __attribute__((aligned(32)));
 
   auto writer = maple::host_command_writer(send_buf, recv_buf);
